@@ -1,4 +1,5 @@
 {
+  "homepage": ".",
   "name": "mcg-site",
   "version": "0.1.0",
   "private": true,
@@ -7,11 +8,13 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "animate.css": "^4.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-markdown": "^9.0.0",
+    "react-router": "^6.17.0",
+    "react-router-dom": "^6.17.0",
+    "react-scripts": "^5.0.1",
     "react-tsparticles": "^1.33.3",
     "web-vitals": "^1.1.2"
   },
@@ -40,7 +43,7 @@
     ]
   },
   "engines": {
-    "node": "16.13.2",
+    "node": "18.12.1",
     "npm": "8.1.2"
   }
 }
